move typescript interfaces, fix selector
