+++ ui/types/vault/services/auth.d.ts
@@ -19,6 +19,12 @@ export interface AuthData {
   mfa_requirement?: MfaRequirementApiResponse;
 }
 
+export interface AuthResponse {
+  namespace: string;
+  token: string; // the name of the token in local storage, not the actual token
+  isRoot: boolean;
+}
+
 export default class AuthService extends Service {
   authData: AuthData;
   currentToken: string;
@@ -30,7 +36,7 @@ export default class AuthService extends Service {
     backend: string;
     data: object;
     selectedAuth: string;
-  }): Promise<any>;
+  }): Promise<AuthResponse>;
   ajax: (
     url: string,
     method: 'GET' | 'POST' | 'PUT' | 'DELETE',
@@ -40,5 +46,6 @@ export default class AuthService extends Service {
       data?: Record<string, unknown>;
     }
   ) => Promise<any>;
+  getAuthType(): string | undefined;
   _parseMfaResponse(mfaResponse: MfaRequirementApiResponse): ParsedMfaRequirement;
 }
